requirements.txt
pymorphy2==0.9.1
gensim==4.3.1
Django==3.2.18
environs==9.5.0
python-telegram-bot==13.13